Ensure each project-dependent action is disabled & hidden as long as PROJECT data key is missing
