+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/toolbar/PullCurrentBranchFastForwardOnlyAction.java
@@ -25,6 +25,7 @@ public class PullCurrentBranchFastForwardOnlyAction extends BasePullBranchFastFo
   @UIEffect
   public void onUpdate(AnActionEvent anActionEvent) {
     super.onUpdate(anActionEvent);
+
     var presentation = anActionEvent.getPresentation();
     if (!presentation.isVisible()) {
       return;
