+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/toolbar/RefreshStatusAction.java
@@ -2,20 +2,23 @@ package com.virtuslab.gitmachete.frontend.actions.toolbar;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.fileEditor.FileDocumentManager;
-import com.intellij.openapi.project.DumbAwareAction;
 import lombok.CustomLog;
 import org.checkerframework.checker.guieffect.qual.UIEffect;
 
-import com.virtuslab.gitmachete.frontend.actions.expectedkeys.IExpectsKeyProject;
+import com.virtuslab.gitmachete.frontend.actions.base.BaseProjectDependentAction;
 import com.virtuslab.logger.IEnhancedLambdaLogger;
 
 @CustomLog
-public class RefreshStatusAction extends DumbAwareAction implements IExpectsKeyProject {
+public class RefreshStatusAction extends BaseProjectDependentAction {
 
   @Override
   public IEnhancedLambdaLogger log() {
     return LOG;
   }
+
+  @Override
+  @UIEffect
+  protected void onUpdate(AnActionEvent anActionEvent) {}
 
   @Override
   @UIEffect
