+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/vcsmenu/OpenMacheteTabAction.java
@@ -3,7 +3,6 @@ package com.virtuslab.gitmachete.frontend.actions.vcsmenu;
 import static com.virtuslab.gitmachete.frontend.resourcebundles.GitMacheteBundle.getString;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.project.DumbAwareAction;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vcs.VcsNotifier;
 import com.intellij.openapi.wm.ToolWindow;
@@ -12,15 +11,18 @@ import com.intellij.openapi.wm.ToolWindowManager;
 import lombok.CustomLog;
 import org.checkerframework.checker.guieffect.qual.UIEffect;
 
-import com.virtuslab.gitmachete.frontend.actions.expectedkeys.IExpectsKeyProject;
+import com.virtuslab.gitmachete.frontend.actions.base.BaseProjectDependentAction;
 import com.virtuslab.logger.IEnhancedLambdaLogger;
 
 @CustomLog
-public class OpenMacheteTabAction extends DumbAwareAction implements IExpectsKeyProject {
+public class OpenMacheteTabAction extends BaseProjectDependentAction {
   @Override
   public IEnhancedLambdaLogger log() {
     return LOG;
   }
+
+  @Override
+  protected void onUpdate(AnActionEvent anActionEvent) {}
 
   @Override
   @UIEffect
