+++ scripts/enforce-call-super-for-overridden-onupdate
@@ -5,9 +5,9 @@ set -e -o pipefail -u
 self_dir=$(cd "$(dirname "$0")" &>/dev/null; pwd -P)
 source "$self_dir"/utils.sh
 
-if git grep --files-with-matches 'void onUpdate' -- '*.java' ':!*/BaseProjectKeyAvailabilityAssuranceAction.java' \
-  | xargs grep --files-without-match 'extends BaseProjectKeyAvailabilityAssuranceAction' \
-  | xargs grep -Pz --files-without-match 'void onUpdate\([\w ]*\) {\n\W*super\.onUpdate'
+if git grep -w --files-with-matches 'void onUpdate' -- '*.java' \
+  | xargs grep -Ew --files-without-match '(class|extends) BaseProjectDependentAction' -- \
+  | xargs grep -Pz --files-without-match 'void onUpdate\([^)]*\) \{\n *super\.onUpdate' --
 then
-  die "You must call 'super.onUpdate' as the first statement if you override 'onUpdate' method and this class does not directly extend 'BaseProjectKeyAvailabilityAssuranceAction'!"
+  die "You must call 'super.onUpdate' as the first statement if you override 'onUpdate' method and this class does not directly extend 'BaseProjectDependentAction'!"
 fi
